Bug fixed, finished cleanup, added all comments
